Arreglos varios en la aplicación, al añadir horas mejorado el sistema de añadir por varios días, se actualiza el gráfico de calendario, botón de añadir funcional y color de CT puesto, añadir, actualizar y quitar horas agregado e icono no funcional para evitar problemas y legenda actualizada para que no pete

+++ AppMovil/composeApp/src/commonMain/kotlin/com/es/appmovil/utils/DTOConverter.kt
@@ -9,7 +9,7 @@ object DTOConverter {
         return TimeCodeDTO(
             idTimeCode = idTimeCode,
             desc = desc,
-            color = color.removePrefix("0x").toULong(16),
+            color = color.removePrefix("0x").toULong(16).toLong(),
             chkProd = chkProd
         )
     }
